+++ ark/type/__tests__/pipe.test.ts
@@ -1,6 +1,14 @@
 import { attest, contextualize } from "@arktype/attest"
-import { assertNodeKind, type Out, type string } from "@arktype/schema"
-import { scope, type Type, type } from "arktype"
+import {
+	assertNodeKind,
+	writeIndiscriminableMorphMessage,
+	writeMorphIntersectionMessage,
+	type MoreThan,
+	type Out,
+	type of,
+	type string
+} from "@arktype/schema"
+import { scope, type, type Type } from "arktype"
 
 contextualize(() => {
 	it("base", () => {
@@ -397,11 +405,34 @@ contextualize(() => {
 	it("discriminable tuple union", () => {
 		const $ = scope({
 			a: () => $.type(["string"]).pipe(s => [...s, "!"]),
-			b: ["boolean"],
+			b: ["number"],
 			c: () => $.type("a|b")
 		})
 		const types = $.export()
-		attest<[boolean] | ((In: [string]) => Out<string[]>)>(types.c.t)
+
+		attest<[number] | ((In: [string]) => Out<string[]>)>(types.c.t)
+		const expectedSerializedMorphs =
+			types.a.raw.assertHasKind("morph").serializedMorphs
+
+		attest(types.c.raw.assertHasKind("union").discriminantJson).snap({
+			kind: "domain",
+			path: ["0"],
+			cases: {
+				'"number"': {
+					sequence: { prefix: ["number"] },
+					proto: "Array",
+					exactLength: 1
+				},
+				'"string"': {
+					in: {
+						sequence: { prefix: ["string"] },
+						proto: "Array",
+						exactLength: 1
+					},
+					morphs: expectedSerializedMorphs
+				}
+			}
+		})
 	})
 
 	it("ArkTypeError not included in return", () => {
@@ -448,120 +479,145 @@ contextualize(() => {
 		attest<Type<(In: string) => Out<number | null | undefined>>>(toMaybeNumber)
 	})
 
-	// TODO: reenable discrimination
-	// it("deep intersection", () => {
-	// 	const types = scope({
-	// 		a: { a: ["number>0", "=>", (data) => data + 1] },
-	// 		b: { a: "1" },
-	// 		c: "a&b"
-	// 	}).export()
-	// 	attest<Type<{ a: (In: 1) => Out<number> }>>(types.c)
-	// 	attest(types.c.json).snap()
-	// })
+	it("deep intersection", () => {
+		const types = scope({
+			a: { a: ["number>0", "=>", data => data + 1] },
+			b: { a: "1" },
+			c: "a&b"
+		}).export()
+		attest<{ a: (In: of<1, MoreThan<0>>) => Out<number> }>(types.c.t)
+		const { serializedMorphs } =
+			types.a.raw.firstReferenceOfKindOrThrow("morph")
 
-	// it("double intersection", () => {
-	// 	attest(() =>
-	// 		scope({
-	// 			a: ["boolean", "=>", (data) => `${data}`],
-	// 			b: ["boolean", "=>", (data) => `${data}!!!`],
-	// 			c: "a&b"
-	// 		}).export()
-	// 	).throws.snap("ParseError: Invalid intersection of morphs")
-	// })
+		attest(types.c.json).snap({
+			required: [
+				{ key: "a", value: { in: { unit: 1 }, morphs: serializedMorphs } }
+			],
+			domain: "object"
+		})
+	})
 
-	// it("undiscriminated union", () => {
-	// 	// TODO: fix
-	// 	// attest(() => {
-	// 	// 	scope({
-	// 	// 		a: ["/.*/", "=>", (s) => s.trim()],
-	// 	// 		b: "string",
-	// 	// 		c: "a|b"
-	// 	// 	}).export()
-	// 	// }).throws(writeUndiscriminableMorphUnionMessage("/"))
-	// })
+	it("morph intersection", () => {
+		attest(() =>
+			scope({
+				a: ["string", "=>", data => `${data}`],
+				b: ["string", "=>", data => `${data}!!!`],
+				c: "a&b"
+			}).export()
+		).throws(
+			writeMorphIntersectionMessage(
+				"(In: string) => Out<unknown>",
+				"(In: string) => Out<unknown>"
+			)
+		)
+	})
 
-	// it("deep double intersection", () => {
-	// 	attest(() => {
-	// 		scope({
-	// 			a: { a: ["boolean", "=>", (data) => `${data}`] },
-	// 			b: { a: ["boolean", "=>", (data) => `${data}!!!`] },
-	// 			c: "a&b"
-	// 		}).export()
-	// 	}).throws.snap("ParseError: Invalid intersection of morphs")
-	// })
+	it("indiscriminable union", () => {
+		attest(() => {
+			scope({
+				a: ["/.*/", "=>", s => s.trim()],
+				b: "string",
+				c: "a|b"
+			}).export()
+		}).throws(
+			writeIndiscriminableMorphMessage(
+				"(In: string /.*/) => Out<unknown>",
+				"string"
+			)
+		)
+	})
 
-	// it("deep undiscriminated union", () => {
-	// 	attest(() => {
-	// 		scope({
-	// 			a: { a: ["string", "=>", (s) => s.trim()] },
-	// 			b: { a: "'foo'" },
-	// 			c: "a|b"
-	// 		}).export()
-	// 	}).throws(writeUndiscriminableMorphUnionMessage("/"))
-	// })
+	it("deep morph intersection", () => {
+		attest(() => {
+			scope({
+				a: { a: ["number", "=>", data => `${data}`] },
+				b: { a: ["number", "=>", data => `${data}!!!`] },
+				c: "a&b"
+			}).export()
+		}).throws(
+			writeMorphIntersectionMessage(
+				"(In: number) => Out<unknown>",
+				"(In: number) => Out<unknown>"
+			)
+		)
+	})
 
-	// it("deep undiscriminated reference", () => {
-	// 	const $ = scope({
-	// 		a: { a: ["string", "=>", (s) => s.trim()] },
-	// 		b: { a: "boolean" },
-	// 		c: { b: "boolean" }
-	// 	})
-	// 	const t = $.type("a|b")
-	// 	attest<
-	// 		Type<
-	// 			| {
-	// 					a: (In: string) => Out<string>
-	// 			  }
-	// 			| {
-	// 					a: boolean
-	// 			  }
-	// 		>
-	// 	>(t)
+	it("deep indiscriminable", () => {
+		const $ = scope({
+			a: { foo: ["string", "=>", s => s.trim()] },
+			b: { foo: "symbol" },
+			c: { bar: "symbol" }
+		})
 
-	// 	attest(() => {
-	// 		scope({
-	// 			a: { a: ["string", "=>", (s) => s.trim()] },
-	// 			b: { b: "boolean" },
-	// 			c: "a|b"
-	// 		}).export()
-	// 	}).throws(writeUndiscriminableMorphUnionMessage("/"))
-	// })
+		// this is fine as a | b can be discriminated via foo
+		const t = $.type("a|b")
+		attest<
+			| {
+					foo: (In: string) => Out<string>
+			  }
+			| {
+					foo: symbol
+			  }
+		>(t.t)
 
-	// it("array double intersection", () => {
-	// 	// attest(() => {
-	// 	// 	scope({
-	// 	// 		a: { a: ["number>0", "=>", (data) => data + 1] },
-	// 	// 		b: { a: ["number>0", "=>", (data) => data + 2] },
-	// 	// 		c: "a[]&b[]"
-	// 	// 	}).export()
-	// 	// }).throws(
-	// 	// 	"At [index]/a: Intersection of morphs results in an unsatisfiable type"
-	// 	// )
-	// })
+		attest(() => $.type("a|c")).throws(
+			writeIndiscriminableMorphMessage(
+				"{ foo: (In: string) => Out<unknown> }",
+				"{ bar: symbol }"
+			)
+		)
+	})
 
-	// it("undiscriminated morph at path", () => {
-	// 	attest(() => {
-	// 		scope({
-	// 			a: { a: ["string", "=>", (s) => s.trim()] },
-	// 			b: { b: "boolean" },
-	// 			c: { key: "a|b" }
-	// 		}).export()
-	// 	}).throws(writeUndiscriminableMorphUnionMessage("key"))
-	// })
+	it("array double intersection", () => {
+		attest(() => {
+			scope({
+				a: { a: ["number>0", "=>", data => data + 1] },
+				b: { a: ["number>0", "=>", data => data + 2] },
+				c: "a[]&b[]"
+			}).export()
+		}).throws(
+			writeMorphIntersectionMessage(
+				"(In: number >0) => Out<unknown>",
+				"(In: number >0) => Out<unknown>"
+			)
+		)
+	})
 
-	// it("helper morph intersection", () => {
-	// 	attest(() =>
-	// 		type("string")
-	// 			.morph((s) => s.length)
-	// 			.and(type("string").morph((s) => s.length))
-	// 	).throws("Intersection of morphs results in an unsatisfiable type")
-	// })
+	it("undiscriminated morph at path", () => {
+		attest(() => {
+			scope({
+				a: { a: ["string", "=>", s => s.trim()] },
+				b: { b: "bigint" },
+				c: { key: "a|b" }
+			}).export()
+		}).throws(
+			writeIndiscriminableMorphMessage(
+				"{ a: (In: string) => Out<unknown> }",
+				"{ b: bigint }"
+			)
+		)
+	})
 
-	// it("union helper undiscriminated", () => {
-	// 	attest(() =>
-	// 		type("string")
-	// 			.morph((s) => s.length)
-	// 			.or("'foo'")
-	// 	).throws(writeUndiscriminableMorphUnionMessage("/"))
-	// })
+	it("helper morph intersection", () => {
+		attest(() =>
+			type("string")
+				.pipe(s => s.length)
+				.and(type("string").pipe(s => s.length))
+		).throws(
+			writeMorphIntersectionMessage(
+				"(In: string) => Out<unknown>",
+				"(In: string) => Out<unknown>"
+			)
+		)
+	})
+
+	it("union helper undiscriminated", () => {
+		attest(() =>
+			type("string")
+				.pipe(s => s.length)
+				.or("'foo'")
+		).throws(
+			writeIndiscriminableMorphMessage("(In: string) => Out<unknown>", '"foo"')
+		)
+	})
 })